requirements.txt
pyYAML
kafka-python